add to watchlist functional; watchlist page implemented; nested navigation to watchlist & watched movies implemented

+++ app/src/main/java/com/rng350/mediatracker/networking/movie/MovieSchema.kt
@@ -17,7 +17,7 @@ data class MovieSchema(
 ) {
     fun toMovieForDisplay(): MovieForDisplay {
         return MovieForDisplay(
-            movieId = this.id,
+            movieId = this.id.toInt(),
             movieTitle = this.title,
             movieOriginalTitle = if (!this.originalTitle.isNullOrEmpty()) this.originalTitle else null,
             moviePremise = this.overview ?: "",
